feat: integrate device-registry in gateway and docker compose

+++ device-registry-service/src/main/java/com/ezhome/deviceregistry/controller/DeviceController.java
@@ -1,7 +1,6 @@
 package com.ezhome.deviceregistry.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ezhome.deviceregistry.dto.AddDeviceRequestDTO;
@@ -31,28 +30,26 @@ public class DeviceController {
         this.deviceService = deviceService;
     }
 
-    //@RequestHeader("X-User-Id") String userId, @RequestHeader("X-User-Role") String role,
-
     @PostMapping("/add")
-    public DeviceDTO addDevice(@Valid @RequestBody AddDeviceRequestDTO req) {
+    public DeviceDTO addDevice(@RequestHeader("X-User-Id") String userId, @RequestHeader("X-User-Role") String role, @Valid @RequestBody AddDeviceRequestDTO req) {
 
-        //verifyAccess(userId, role);
+        verifyAccess(userId, role);
         return deviceService.addToRegistry(req);
         
     }
 
     @GetMapping("/find/{id}")
-    public DeviceDTO findDevice(@Valid @PathVariable String id) {
+    public DeviceDTO findDevice(@RequestHeader("X-User-Id") String userId, @RequestHeader("X-User-Role") String role, @Valid @PathVariable String id) {
 
-        //verifyAccess(userId, role);
+        verifyAccess(userId, role);
         return deviceService.findinRegistry(id);
         
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteDevice(@Valid @PathVariable String id) {
+    public void deleteDevice(@RequestHeader("X-User-Id") String userId, @RequestHeader("X-User-Role") String role, @Valid @PathVariable String id) {
 
-        //verifyAccess(userId, role);
+        verifyAccess(userId, role);
         deviceService.deleteFromRegistry(id);
         
     }
@@ -60,7 +57,7 @@ public class DeviceController {
     private void verifyAccess(String userId, String role) {
         UUID.fromString(userId);
 
-        if(role != "ADMIN"){
+        if(!role.equalsIgnoreCase("ADMIN")){
             throw new AccessException("Account does not have permission");
         }
     }
